Modified C# extensions to only facilitate value tuple retuns. Removed Validate, since it ended up as a duplicate of IsValid. Updated documentation
