+++ DotNet/SSNDK.CSharp.Test/IsValidTest.cs
@@ -35,5 +35,30 @@ namespace SSNDKCSharp.Test
             var result = !success && error == ErrorReason.NonDashCharacter;
             return result.ToProperty();
         }
+
+        [Property]
+        [Trait("Category", "Unit")]
+        public Property IsValid_works(DateTimeOffset x,
+            bool dash, NonNegativeInt controlCode, bool useModula11Check)
+        {
+            var sut = GetSSN(x, dash, controlCode);
+
+            var (success, error) = sut.IsValid(useModula11Check);
+            if (success)
+                return true.ToProperty();
+            else
+            {
+                if (useModula11Check)
+                {
+                    var indices = Enumerable.Range(0, 10)
+                        .Select(n => n < 6 ? n : (dash ? n + 1 : n))
+                        .ToArray();
+                    var isModula11 = SSNDK.Helpers.sumOfProduct(indices, sut) % 11 == 0;
+                    if (isModula11) return false.ToProperty();
+                    else return (error == ErrorReason.Modula11CheckFail).ToProperty();
+                }
+                return false.ToProperty();
+            }
+        }
     }
 }
